requirements.txt
# Core Dependencies
python-arango>=7.5.0
python-dotenv>=1.0.0
requests>=2.31.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Data Processing
numpy>=1.24.0

# Optional: For notebook support
jupyter>=1.0.0
ipykernel>=6.25.0

# Arango Entity Resolution Library (Official PyPI Release)
arango-entity-resolution==3.1.0

# GraphRAG Dependencies
docling==2.26.0
pymupdf4llm
PyMuPDF
pymupdf_layout
markdownify
aiohttp

